rename icon and add to markdown editor
