+++ src-docs/src/views/icon/icons.js
@@ -132,7 +132,6 @@ export const iconTypes = [
   'link',
   'list',
   'listAdd',
-  'listOptions',
   'lock',
   'lockOpen',
   'logstashFilter',
