+++ src/components/icon/icon.tsx
@@ -104,6 +104,7 @@ const typeToPathMap = {
   editorAlignLeft: 'editor_align_left',
   editorAlignRight: 'editor_align_right',
   editorBold: 'editor_bold',
+  editorChecklist: 'editor_checklist',
   editorCodeBlock: 'editor_code_block',
   editorComment: 'editor_comment',
   editorDistributeHorizontal: 'editorDistributeHorizontal',
@@ -202,7 +203,6 @@ const typeToPathMap = {
   link: 'link',
   list: 'list',
   listAdd: 'list_add',
-  listOptions: 'list_options',
   lock: 'lock',
   lockOpen: 'lockOpen',
   logoAWS: 'logo_aws',
